requirements.txt
django==4.2.10
django-crispy-forms==2.0
crispy-bootstrap5==0.7
djongo==1.3.6
pymongo==4.6.1
pillow==10.1.0
python-dotenv==1.0.0
django-storages==1.14.2
django-allauth==0.57.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0